docs: use patch command to edit ConfigMap (#1498)
Inspired by https://www.youtube.com/watch?v=eFawBTtoJmE to use
a one liner to update the ConfigMap. This makes it easier to
run and automated as opposed to the `kubectl edit` command.

Also removes `$` from command blocks and uses the `bash` markdown
syntax for consistency, which is an alias for `shell, sh, zsh`.