pkg/*: remove type assertions for cache items that can be safely asserted
Previously we added type assertions for kubernetes resources in their
caches. These type assertions can only result from bugs in the code
if there is a mismatch between the resource type in the informer
and the code that asserts for the resource type in the informer cache.

Such a scenario is not possible to test if the code is correct (as is
the case now), and the code must be buggy for a potential unit test
to even be able to test such mismatches.

This change removes such checks in the informer caches since its not
possible to even unit test the need for such code. It also simplifies
the return values of some functions that were only returning an error
for these scenarios.