envoy/cds: use downstream connection for HTTP protocol selection
Currently OSM does not honor the protocol version of downstream
connections. As a result, HTTP2 clients are unable to route traffic
to HTTP2 servers. By default, Envoy uses HTTP 1.1 if
http2_protocol_options is not configured, as is the case now.

This change configures Envoy to use an HTTP protocol version
based on the downstream connection, allowing both HTTP1.1 and
HTTP2 clients to be able to route traffic. Default values
are used for the http2_protocol_options, but option must be
explicitly set as a part of the cluster config.

This change is primarily useful for HTTP2 connections over
plaintext, ex. with grpc.WithInsecure().

Testing done:
```
 curl -v --http2 client-app: bookbuyer" \
    -H "user-agent: Go-http-client/1.1" \
    http://bookstore.bookstore/books-bought

< HTTP/1.1 200 OK
< booksbought: 99
< identity: bookstore-v2
< date: Wed, 19 Aug 2020 18:40:00 GMT
< content-length: 1397
< content-type: text/html; charset=utf-8
< x-envoy-upstream-service-time: 2
< server: envo
```

Resolves #1589

Signed-off-by: Shashank Ram <shashank08@gmail.com>