Replacing Zipkin -> Jaeger for tracing
Note: Zipkin could refer both the app and the protocol format
for tracing.

For things referring to the app specifically, this commit replaces
zipking with jaeger in all related forms.

For OSM, and to keep the code as agnostic as possible,
the ZipkinCluster has become the TracingCluster (as anything
could be listening to tracing in Zipkin protocol), and the zipkin
endpoint configuration remains Zipkin, as the expected listener
is still expected to listen in Zipkin format.