Makefile: use long git SHA for consistency
We publish images with long form git SHA as image tags, and use
the long from as defaults elsewhere, so use the long form SHA
for the CLI version as well. This is useful to compare a CLI
version to its image tag, to ensure compatibility.

Also removes an unused script that is no longer required.

Signed-off-by: Shashank Ram <shashank08@gmail.com>

Partly addresses ambiguity seen in #1527