install: make OSM ImagePullPolicy configurable
Commit a42c772 changed the default pull policy for
OSM images to `IfNotPresent`, which breaks local
testing where the same CTR_TAG is used across
reruns of the demo. As a result, when a new change
is made locally and the demo is rerun, they don't
get picked by the demo. Even using the git SHA as the
default tag will not solve this problem because an
amended commit will have the same image tag.

This change makes the image pull policy configurable
in the install cli, with the default being set
to `IfNotPresent` as before. The demo is updated
to set the pull policy to `Always` so that rerunning
the demo always picks up the latest built image,
irrespective of the tag.

Resolves #1605

Signed-off-by: Shashank Ram <shashank08@gmail.com>