Rename Route type to HTTPRoute where applicable (#1612)
Currently the code refers to an HTTPRoute as a Route in a traffic target
policy. Rename this to HTTPRoute to be explicit. In the future, a route
for different traffic types will be supported, ex. TCPRoute.

This also helps disambiguate the HTTP route type in the traffic target from
the XDS Route type in Envoy.

Signed-off-by: Shashank Ram <shashank08@gmail.com>