ref(build): shrink docker build contexts
Currently, each Docker image is built using the root of the repository
as the root of its build context. When binaries have already been built,
the total size of the build context is around 160MB which takes a
noticeable amount of time for the Docker daemon to copy, especially when the
build cache is warm. This change scopes the Docker build context for
each image to be as small as possible by separating the files used to
build each image into different directories.

Here are some rough benchmarks showing the speedup.

Before

    make docker-build  11.84s user 8.27s system 79% cpu 25.317 total

After

    make docker-build  10.14s user 5.71s system 166% cpu 9.536 total

Note that you may have to manually remove the demo/bin and bin/
directories after pulling this change to avoid errors building binaries.