sidecar-injection: provide granular control for injection
Currently it is not possible to enroll specific pods within
the mesh for sidecar injection without enrolling the entire
namespace. This behavior is not desirable when several
application pods run in the same namespace and only certain
pods need to be enrolled in the mesh with a sidecar.

This change decouples the sidecar injection for pods from
the namespace monitoring workflow.

A pod will be injected with the sidecar if its namespace
is monitored, and either of the following are true:

1. The pod is explicitly enabled for injection
2. The namespace is enabled for injection and the pod
   is not explicitly disabled for injection

This change also extends the `osm namespace add` command
with an `--enable-sidecar-injection` flag which automatically
adds the sidecar injection annotation on the namespace.
Similarly, the `osm namespace remove` command is updated
to remove the annotation if present.

It also fixes an existing bug in the `osm namespace remove` command
where all the labels are removed instead of just the monitor label.
Additionally, tests are updated to check for the labels and
annotations as expected.

Resolves #1617

Signed-off-by: Shashank Ram <shashank08@gmail.com>