catalog: GetServiceEndpoints API
We are introducing a new API called GetServiceEndpoints, which will
attempt to return the virtual or cluster IP endpoint for a given service,
or alternatively return all the individual instance endpoints at which
the service is avilable at.

To properly test, needs some larger changes that I will follow up
with on a commit on top.