ref(ci): Use default KUBECONFIG handling
The clientcmd package in k8s.io/client-go provides helpers for getting a
kube config that automatically check in-cluster first and fallback to
out-of-cluster and handle parsing and merging files listed in the
`$KUBECONFIG` variable to match the behavior of `kubectl`. This change
eliminates the need to explicitly define `$KUBECONFIG` for the maestro
when the file is in the default location.