kubernetes/events: introduce k8s event recording
This change introduces a kubernetes event recording
framework. The events framework allows recording
kubernetes events on a specific k8s object using
the `events.NewEventRecorder()` api, or using a generic
event recorder using `events.GenericEventRecorder()`.
The generic event recorder is a singleton that is
used to record events against the osm-controller pod
resource.

As a start, some of the fatal events in controller's
startup have been updated to use the event recording
api.

Part of #654

Signed-off-by: Shashank Ram <shashank08@gmail.com>