mock/NamespaceController: rename file and update comment
After e7f383f84b7e596180b86db8b6b474a504f21e69, the Controller
interface in the `namespace` pkg was renamed to `NamespaceController`
in the `kubernetes` pkg. Since the mock is specific to the
`NamespaceController` interface, this change renames the mock
and also updates the comment in the file to avoid linting
errors.

Signed-off-by: Shashank Ram <shashank08@gmail.com>