Resolve ambiguity between webhook config and webhook name
Currently the code uses the term 'webhook name' or the variable
'webhookName' when in fact it is referring to the
mutatingwebhookconfiguration name. This is confusing because
the mutatingwebhookconfiguration has a webhook with a name, and
it would be easy to confuse the two.

This change resolves this ambiguity. Commit 03e00587 removed
part of this ambigutiy by renaming the the variable used
to represent the name of the webhook.

Signed-off-by: Shashank Ram <shashank08@gmail.com>