cmd/certificate: handle error scenarios when fetching secret
This change refactors the `getCertFromKubernetes` function to
handle error scenarios. The function is only expected to error
if it finds a secret with an invalid root certificate. It also
helps remove the call to `log.Fatal` while parsing the root
certificate.

Also updates the existing test context to test error scenarios.

Signed-off-by: Shashank Ram <shashank08@gmail.com>