certificates: allow reconfiguring cert validity for short lived certs
This change allows reconfiguring the validity period for short lived
certs. The validity period of such certs are no longer static, but
dynamically retrieved from the osm-config ConfigMap.
Additionally, the validity period at the time of certificate issuance
must be explicitly be specified to prevent issuing certs with wrong
validity.

Part of #1757

Signed-off-by: Shashank Ram <shashank08@gmail.com>