kubernetes/events: remove unnecessary logging
Each of the xxxEvent() apis exposed by the EventRecorder already
log at the corresponding log levels. This change removes the
additional trace logging which results in duplicate log messages
for each event.

The additional logging as a part of the event recorder also
results in data races in unit tests.

Part of #1791