tests/e2e: Increasing clenaup time
With enough test applications/namespaces in combination with a slow
instance of gh actions, we've seen some delays to clean up all test
namespaces and test failures.

Relaxing further the wait-for-namespaces cleanup time, as re-running
the test suite for a non-osm-related issue is a lot more expensive
compared to grating additional infra cleanup time.