cli: rename global namespace flag (#1887)
This change is an RFC to discuss the proposal
to rename the global namespace flag exposed in
osm cli, which reflects the control plane namespace.
The reason to do so is because local command flags
could also expose a namespace flag, which don't
show up in the help text for the command. Moreover,
the global namespace flag isn't used by several
commands (ex. osm metrics, osm proxy etc.).

The proposal is to rename the global flag to represent
OSM's namespace to `osm-namespace` so that local
commands wanting to expose `namespace` as a flag can
do so with command specific description.

Part of #1884