fix(build): ensure controller version is not a chart version
Currently, we make tags of the form `v.X.Y.Z` for binary releases and
`chart/X.Y.Z` for the Helm chart. The version is passed to the
osm-controller binary via a linker flag. If both a chart and regular
semver tag exist on the same commit, the chart tag may be passed instead
of the regular version tag. This change ensures only our regular semver
tags are used for the versions embedded in the binaries.