catalog: correctly program shorthand service FQDN
Currently, while programming the service FQDNs for an
upstream service in the downstream envoy's virtual_host config
the shorthand service FQDN using the service name without the namespace
is always programmed. This is incorrect because the same service (name)
could exist in multiple namespaces. The service FQDN without the
namespace should only be programmed on downstream clients that are
in the same namespace.

This change addresses this bug and adds tests for the same. It also
removes a duplicated test copied into an unrelated package.

Resolves #951

Signed-off-by: Shashank Ram <shashank08@gmail.com>