charts/osm: Increase max memory limit
Recent findings show OSM to hit early max limits (127MB)
with as much as 120~140 pods.

It's hard to estimate the memory used per envoy as it
will also very much depend on the traffic structrues used
by the deployment (service and trafficspec caches, etc).

On the specific tests, it showed a 0.5~0.6 MB/envoy.