injector: reuse patch helper for patching labels and annotations
Labels and Annotations are both map types, so the same patching
code can be used for both with the patching path being the only
difference. This change reuses the helper patch map types.
The existing tests validate patching map types, additionally
the `create_patch` function which patches labels is also tested
which ensures that patching of labels works as expected.