identity: add abstraction for service identity
Adds a new `ServiceIdentity` type that is to be used to provide
identity to services managed by OSM. Provides an API to use
Kubernetes ServiceAccount as service identities.

The kubernetes based service identity makes use of a trust domain
that determines the domain within which an identity can be trusted.
Within the same cluster, a Kubernetes service identity would be
"service-account.namespace.cluster.local". The trust domain provides
the ability to extend a `ServiceIdentity` beyond kubernetes domains.

Part of #1965