pkg/*: issue service cert directly via cert manager api
This change allows SDS to directly use the IssueCertificate api
implemented by the certificate manager in use, instead of going
through MeshCatalog. This simplifies certificate issuance and code
duplication (such as catalog previously checking cert cache which
the IssueCertificate api implements within itself).

The tests are updated to reflect this change.

Part of #1965