fix(install): clean up resources when install fails
This change adds the behavior of the `--atomic` flag to our invocation
of `helm install` in the `osm install` command. With `--atomic`, Helm
will uninstall the release if it fails to install. `--atomic` also
implies `--wait` which will wait for pods, pod workloads, services,
PVCs, and CRDs to become ready before exiting.

A `--timeout` flag is also added to control the maximum amount of time
to wait for the resources to install and become ready. Its value is
passed through to Helm and behaves exactly the same as the `--timeout`
flag for `helm install`.

Fixes #1530