cli,configurator: reuse api to parse osm-config ConfigMap
Reuses the configurator package to parse the permissive mode
config key in the ConfigMap as a part of osm cli.
Functionally, this change does not change the existing behavior
of the configurator client in osm-controller, which ignores errors
while initializing parsing the keys in the osm-config ConfigMap to
use default values where applicable.

Resolves #1962