configurator: optimize ConfigMap informer filtering (#2000)
This change makes use of the filtering capability in informers
to filter API responses on the server side, by using a FieldSelector
to filter on the name of the ConfigMap as opposed to filtering all
objects on the client (osm-controller) side. The informer is already
namespaced, so the namespace and name should be enough to accurately
only have OSM's ConfigMap in the informer cache and events.

This change will also reduce the number of ConfigMap related events
the controller will receive when there are additional ConfigMaps in
the controller's namespace (ex. Grafana's ConfigMap when Grafana is
deployed as a part of control plane), and thus improving performance
in such scenarios. The informer resyncs are on a per resource basis,
so unnecessary events can take up processing resources.