kubernetes/client: remove announcement channel for namespace informer
The namespace informer is only necessary to check if namespaces are
monitored (via the monitoring label), which is done via the cache
store provided by client-go.

We do not need to trigger events to catalog when a namespace is
is added/updated/deleted for a few reasons:

1. Envoy configuration updates are not concerned with individual
   namespace resources events. Adding a new namespace or updating
   the namespace (ex with a label/annotation) does not concern
   proxy configurations. Deleting a namespace however does affect
   proxy configurations, but since the catalog subscribes to pod
   and service specific events, a namespace deletion event would
   also result in pod/service deletion events in such a namespace
   which will suffice to recompute proxy configs for proxies
   fronting affected pods/services.

2. Because of 1, annoucements specific to the namespace resource are
   not very meaningful to catalog/XDS verticals. As such, the APIs
   to check if namespaces are monitored leverage the cache store and
   don't need to announce events.

3. Triggering announcements for resources that don't affect the proxy
   configuration directly incurs a huge performance overhead, since
   a single announcement will result in configurations being recomputed
   for every single connected proxy. Moreover, as users label/annotate
   their namespaces, the performance will exponentially degrade with
   the number of number of updates on a per namespace basis.

This change also updates the tests that rely on the namespace cache
store to be asynchronously checked. Adding an annoucement just for
the test will result in issues in the controller when there are
no readers for the channel, with the events being blocked and causing
subsequent to be stalled.