kubernetes: add an API to retrieve ServiceAccounts for a Service
Adds an api to the kubernetes.Controller interface to return the
list of all ServiceAccounts associated with a given service. A
service backed by multiple pods belonging to different ServiceAccounts
is associated with ServiceAccounts, and the api will return a unique
list of associated ServiceAccounts.

An example usage of this api will be when client proxies need to
validate upstream services based on SAN matching in certificates,
which are based on ServiceAccounts.