tests(debug server): fix race condition
Currently, the tests that dynamically start/stop the debug HTTP server
rely on a `sync.WaitGroup` that signals when a mocked debug server has
been started or stopped. `TestConfigureDebugServerErr` was failing
occasionally because `configureDebugServer` wasn't guaranteed to finish
stopping the server before the test's `sync.WaitGroup` signalled it was
done.

This change removes the `sync.WaitGroup` from the tests and instead adds
a new channel parameter to `configureDebugServer` that will emit errors
(or nil if successful) whenever the server is started or stopped. Then,
the tests can be sure that `configureDebugServer` has finished before
checking any values.