envoy: correctly validate SANs in certificate (#2021)
This change fixes how a client validates the service certificate presented
by an upstream server. Previously, the identity of the upstream
server was being validated against a list of upstream service
identities that the client is allowed to communicate to. This is
vulnerable to security issues where a server not authorized
to run the given upstream service the client is attempting to reach
can successfully trick the client as long as the client is allowed
to reach the unauthorized service identity for a different service.

This change prevents this by ensuring the service identities presented
by the upstream server are indeed authorized to run the service the
client is attempting to communicate to.

This change also decouples the context of validation context for an
SDS secret from the service certificate. Previously, the service
cert was used as the validation context as well.

With this change, certificate validation happens as follows.

1. Outbound
- Upstream cluster's UpstreamTlsContext points to the client's service
  cert, with validation context pointing to the upstream server root cert.

- During TLS handshake, client presents the client service cert, and
  validates server cert against the root cert validation context
  specified in the upstream cluster's UpstreamTlsContext. The upstream
  cert's SANs are matched against service identities corresponding to
  upstream service.

2. Inbound
- Inbound cluster's DownstreamTlsContext points to the local service
  cert, with validation context pointing to the local server's root
  cert.

- During TLS handshake, downstream presents its cert to the local
  cluster, and the local cluster's DownstreamTlsContext validates
  that the service identity presented by the downstream is in the
  list of allowed service identities that can communicate with the
  local (or upstream) cluster. This can be made stricter on a per
  client basis but needs a DownstreamTlsContext per client, which
  is a separate effort in intself.

This change also updates the SDS test to dynamically update mocks
for each test in the table driven test suite.

Resolves #2024