ref(injector): pass pod object to getEnvoySidecarContainerSpec
This change refactors `getEnvoySidecarContainerSpec` to accept the pod
object as an argument so its other fields do not have to be passed as
additional arguments explicitly. In particular, the pod's owner
references will be read to keep track of which workload it is a part of,
which is needed to implement SMI metrics.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>