envoy/rbac: add generic RBAC policy builder
This change adds a new rbac pkg to build RBAC policies.
Currently this change doesn't affect existing functionality
and the goal is to have a generic rbac policy builder
that can be used when an RBAC policy needs to be built.

LDS builds RBAC policies to enforce RBAC on inbound
connections from downstreams. RDS will require to enable
RBAC on a per route basis in the future. This pkg will
serve as a modeler for RBAC policies.

Signed-off-by: Shashank Ram <shashank08@gmail.com>