envoy/lds: update error handling while building HTTP filter chain
To be able to construct the outbound filter chain for an upstream
service, the upstrea service needs to have endpoints. If endpoints
are missing, then the filter chain cannot be constructed.

This change logs an error in the case when the filter chain for
an upstream cannot be constructed. It also continues building
the filter chains for other services so that connectivity to
other upstream services are not impacted when a particular
upstream filter chain cannot be constructed.

Signed-off-by: Shashank Ram <shashank08@gmail.com>