prometheus: drop time-bucket series, increase ram limit to 2G
(tentative)

We are using a large amounts of labels to tag our metrics with addtional
dimensions of interest. This is good for visibility, however this explodes
in RAM consumption on Prometheus once scraped.

Most of the heavily unscaling metrics are the time-bucket series, which
already have large number of dimensions due to the buckets themselves.

- By filtering out time-bucket series (which might contain latency and
request total time buckets), we can cut the memory consumption around
66~70%, allowing larger scales to work under 2Gb, compared with same
scale crashing at 4Gb).
- Additionally, rose the RAM limit for prometheus to 2Gb.

Purely tentative, proper discussion on relabeling has to happen, this
is however required to even support slightly larger number of pods for
the time being.

Signed-off-by: Eduard Serra <eduser25@gmail.com>