endpoint/providers: add api to retrieve port:protocol mapping for service
Adds an api to retrieve the port to protocol mapping for a service.
If an `AppProtocol` is unspecfied, it will default to `http` in lieu
of current OSM protocol support.
This will be used later on to be able to correctly filter inbound
traffic arriving on a port, as described in commit ff54bd77.

This change also refactors the tests which were relying on objects
created by previous tests causing test pollution, and some additional
cleanup.

Part of #1521

Signed-off-by: Shashank Ram <shashank08@gmail.com>