tests/framework: Add GetTestFile (#2136)
We might be interested in near future to save test results, data or logs on disk.
This commit introduces a helper for tests to prefix filenames and create test folder
upon use.

Added a random uin64 identifier for a test, which will be used
to identify a certain test run. Note these are random, but not unique.

GetTestFile is fed with a filename and returns a path to a file
inside test folder. If a folder for current test does not exist, it is
created in the process.

Signed-off-by: Eduard Serra <eduser25@gmail.com>