demo/tcp: adds simple tcp client server demo
This change adds simple tcp client and server apps to
demo tcp support in OSM. The client opens a connection,
sends a number of messages, and prints the server response.
The server simply echoes back the client's message.

Part of #1521, to be used as a simple demo to showcase TCP
traffic when OSM supports TCP. TCP support is still in progress
but the apps don't depend on it.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>