Grafana/dashboards: change rate to irate
For display purposes, and for quickly changing values, `rate` properly
gives an average but this gets diluted if multiple values with
high variance appear in the same bucket.

For measures which have high variance and might be of interest
to understand the deltas with as much precision possible,
irate will always compute the delta between previous and current point,
regardless of bucket size, which makes sense to understand sudden
changes of trend which could get diluted/avgd otherwise.