doc(charts): doc why failurePolicy must be Fail
failurePolicy in the MutatingWebhookConfiguration should
always be set to Fail so that no new resources get created
and thus bypass SMI Traffic Target policies if the for some
reason the webhook server is down. This is intentional and
guarentees that new pods in the monitored namespaces and those
intended to be injected with the sidecar are indeed injected.
Setting this value to Ignore may expose Pods to traffic
they should not be exposed to.

Signed-off-by: Michelle Noorali <michellemolu@gmail.com>