tests/scale: first scaling test
Body and operation of the test is close to identical
to existing `e2e_trafficsplit_test.go`, with few
differences to be able to iterate repeatitively over
the body of the test itself.

Test will repeat the test operation till a failure is
observed. At each iteration, the timestamps are recorded,
and later statistics for relevant pods are gathered and
output as part of the results, for every iteration and its
duration. See below additional examples in comments.