endpoint/kube: Move caches to kubecontroller layer
This commit removes a repeated cache (pods) and moves the endpoint cache
to kubecontroller in an attempt to unify code and the interfaces
that handles kubernetes objects.

This commit also removes the mocks on provider tests that were trying
to dynamically simulate kubecontroller logic (which is highly error-prone),
in favour of using the real kubecontroller object instead. Specific information
for each category has been added on their respective ginkgo `Defines`.

- Pod cache from provider has been removed. Kubecontroller held one
all along already.
- Endpoint cache has been moved from provider/kube to Kubecontroller.
- API to get endpoints for a service was added (GetEndpoints).
- Removed the Announcement channel for provider's interface and implementation,
  subscription/announcement capabilities are still available through pubsub.

Tests:
- Moved the tests that were using dynamic mocks to use a real kubecontroller object
instead.
- Added some more synchronization as there were still some races lurking around.