tests/e2e: add e2e test for TCP traffic (#2194)
This change adds an e2e test for TCP traffic. The test verifies
that the inbound and outbound TCP filtering and proxying works
as expected. It uses `netcat` to initiate TCP connections
to the `tcp-echo-server` that echoes back the response.

Due to the absence of TCPRoute spec support at this moment, the
test uses a TrafficTarget without specific TCP rules to verify
inbound and outbound proxying of TCP traffic. The test will need
to be updated to be able to perform RBAC on TCP routes when the
TCPRoute spec is available to use.

It also updates the client container image for the test to rely
on an image with additional tools.

Part of #1521

Signed-off-by: Shashank Ram <shashr2204@gmail.com>