docs(release): clarify main branch version update
This change more generally describes the final step of the release
process which updates the main branch with the new version references.
For patch releases, merging the release branch into main will
unnecessarily add commits cherry-picked from main to the release branch
a second time, making the history on main harder to follow.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>