envoy/ads/stream.go: Force resync of an envoy when it connects
On a controller re/start, the controller synchronize all kubernetes
objects and only after starts the ADS server.

When an envoy reconnects to the new ctrl plane, there might be no event
or trigger that pushes a new configuration to the envoy, which in the case
of a restart, new service certificates need to be issued and pusghed.

For the time being, we will mandatorily push a full config on a
connecting envoy, which will guarantee it gets the new certs in case
of a restart.

Fixes: #2131