ci: only run maestro once for tresor
Previously we were using maestro for integration tests and testing
different combinations of the product via maestro. All of these
have been moved to the e2e testing suite and are implemented as
separate tests. This change removes the duplicate tests from
maestro and only runs demo integration test with Tresor to ensure
maestro is not broken and the automated demo continues to work.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>