catalog: update api to retrieve traffic targets (#2222)
Updates the api added in 85f8810a to encode the fully qualified
service identities instead of the identities being encoded
only as the service account information. This avoids leaking
service account primitives to XDS components that will leverage
this API. This API will be used to build the RBAC policies and
will not have to deal with service accounts after this change
due to the ServiceIdentity abstraction.

This change additionally adds a workaround to implement TCP route
matching by relying on a `ports` label on the v1alpha3 TCPRoute
resource which does not have a spec yet. Upon migration to v1alpha4
of TCPRoute, the logic to rely on labels will be removed.

Part of #1521

Signed-off-by: Shashank Ram <shashr2204@gmail.com>