metricsstore: integrate metrics store framework
This change integrates the metrics store framework
to allow osm-controller to publish metrics exposed
on the http server's `/metrics` path.

It introduces a `DefaultMetricsStore` that can be
used throughout the controller components. The metrics
store now publishes metrics on k8s api events.
It also updates the metrics to follow best practices
on constant labels to avoid setting the same label on
all metrics.

Part of #902

Signed-off-by: Shashank Ram <shashr2204@gmail.com>