prometheus: limit scrapping to currently used metrics only
Prometheus is still (by default) scrapping a large number of metrics we
are currently not using on our deployments.

This does not mean to say they are usless (which they are not), but by
default adding all metrics is causing serious memory bottlnecks when
increasin scale even at lower numbers.

A previous commit hacked the bucket timeseries out of the equation, as they
amounted for roughly 75% of memory usage on prometheus (from 20Mb per pod to 6Mb)
https://github.com/openservicemesh/osm/pull/2117

This patch pinpoints the exact metrics we use on our deployments, and limits
ingestion to only metrics we we are currently using in our current.
This brings down 6Mb to 2Mb per pod, further allowing operation of prometheus under
larger setups.

Fixes: #2122
Signed-off-by: Eduard Serra <eduser25@gmail.com>