envoy/lds: build outound service filter chains per port
This change updates how the outbound filter chains are built
per upstream service on the downstream proxy. Instead of
relying on the http_inspector listner_filter to inspect
the application protocol, filter chains use the destination
port of the service as a match, and the type of filter (http,
tcp etc.) is determined by the application protocol for that
port. The port-protocol mapping is already leveraged for
protocol specific filtering and proxying on the inbound
listener, this is now being leveraged for outbound filtering
as well.

This is also beneficial for the following reasons:
1. Application protocols such as mySQL where the server
   initiates the application protocol handshake will result
   in listener filters such as tls_inspector and
   http_inspector to time out since it will wait for data
   from the client. With the removal of the http_inspector
   for protocol based filtering, this problem is avoided.

2. By leveraging the appProtocol specified on a port, new
   application protocols can be easily supported via
   corresponding Envoy filters. Ex. an appProtocol=mySQL
   or appProtocol=gRPC can be used to determine the type
   of filter to leverage. This will not be possible
   with using the http_inspector alone as before.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>